Implement separate module reloading
Now takes the module name from the file name, instead of from the class instance. Modules can be loaded/unloaded and enabled/disabled individually, on the fly.

Event handlers should be attached in the new init() hook on modules. This will be called when the module is enabled.

Adds commands for enabling/disabling and reloading modules to the System module.

(Some of it is a bit of a mess right now, though!)