rework app.js: get rid of readline-based command line interface
Because we didn't use the CLI anyway.

Now shuts down more gracefully, logging errors and everything.
`forever` will restart the process for us, so we can just exit
with a nonzero exit code and all will be well. We won't have
all the prompts anymore in log files either, which is nice.